fix: move clientCommandMessageReg to utils
